refactor: Move school schema logic to '`schoolSchema.repository.ts`'
