+++ src/database/repositories/index.ts
@@ -1,2 +1,3 @@
 export * as SchoolRepository from "./school.repository";
+export * as SchoolSchemaRepository from "./schoolSchema.repository";
 export * as StudentRespository from "./student.repository";
